+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v0.11.0 (2023-10-18)
 
 * Update dependency to use a version without a bug
 * If the output file already exists, then exit with error so script can know